backend/requirements.txt
django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.28.0
opencv-python-headless==4.5.5.64
pillow==9.3.0
numpy==1.23.4
django-environ==0.11.2
scikit-learn==1.3.2
django-cors-headers==4.4.0
